Move APIKeyValidator to purchases module
