+++ purchases/src/main/kotlin/com/revenuecat/purchases/Purchases.kt
@@ -19,7 +19,6 @@ import com.android.billingclient.api.BillingClient
 import com.android.billingclient.api.BillingClientStateListener
 import com.android.billingclient.api.BillingResult
 import com.android.billingclient.api.SkuDetails
-import com.revenuecat.purchases.common.APIKeyValidator
 import com.revenuecat.purchases.common.AppConfig
 import com.revenuecat.purchases.common.Backend
 import com.revenuecat.purchases.common.BillingAbstract
